+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/config/CustomKubernetesResource.java
@@ -15,13 +15,13 @@
  */
 package com.netflix.spinnaker.clouddriver.kubernetes.config;
 
-import com.netflix.spinnaker.clouddriver.kubernetes.v2.description.KubernetesSpinnakerKindMap;
+import com.netflix.spinnaker.clouddriver.kubernetes.description.SpinnakerKind;
 import lombok.Data;
 
 @Data
 public class CustomKubernetesResource {
   private String kubernetesKind;
-  private String spinnakerKind = KubernetesSpinnakerKindMap.SpinnakerKind.UNCLASSIFIED.toString();
+  private String spinnakerKind = SpinnakerKind.UNCLASSIFIED.toString();
   private String deployPriority = "100";
   private boolean versioned = false;
   private boolean namespaced = true;
