+++ clouddriver-kubernetes/src/test/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v2/description/KubernetesSpinnakerKindMapSpec.groovy
@@ -17,6 +17,7 @@
 package com.netflix.spinnaker.clouddriver.kubernetes.v2.description
 
 import com.google.common.collect.ImmutableSet
+import com.netflix.spinnaker.clouddriver.kubernetes.description.SpinnakerKind
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.description.manifest.KubernetesKind
 import com.netflix.spinnaker.clouddriver.kubernetes.v2.op.handler.KubernetesHandler
 import spock.lang.Specification
@@ -25,7 +26,7 @@ class KubernetesSpinnakerKindMapSpec extends Specification {
   void "the kind map is properly initialized"() {
     given:
     def mockHandler = Mock(KubernetesHandler) {
-      spinnakerKind() >> KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
+      spinnakerKind() >> SpinnakerKind.INSTANCES
       kind() >> KubernetesKind.REPLICA_SET
     }
 
@@ -33,32 +34,32 @@ class KubernetesSpinnakerKindMapSpec extends Specification {
     def kindMap = new KubernetesSpinnakerKindMap([mockHandler])
 
     then:
-    kindMap.translateSpinnakerKind(KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES) == ImmutableSet.of(KubernetesKind.REPLICA_SET)
-    kindMap.translateKubernetesKind(KubernetesKind.REPLICA_SET) == KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
+    kindMap.translateSpinnakerKind(SpinnakerKind.INSTANCES) == ImmutableSet.of(KubernetesKind.REPLICA_SET)
+    kindMap.translateKubernetesKind(KubernetesKind.REPLICA_SET) == SpinnakerKind.INSTANCES
   }
 
   void "the kind map properly groups kinds"() {
     when:
     def kindMap = new KubernetesSpinnakerKindMap([
       Mock(KubernetesHandler) {
-        spinnakerKind() >> KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
+        spinnakerKind() >> SpinnakerKind.INSTANCES
         kind() >> KubernetesKind.REPLICA_SET
       },
       Mock(KubernetesHandler) {
-        spinnakerKind() >> KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
+        spinnakerKind() >> SpinnakerKind.INSTANCES
         kind() >> KubernetesKind.DEPLOYMENT
       },
       Mock(KubernetesHandler) {
-        spinnakerKind() >> KubernetesSpinnakerKindMap.SpinnakerKind.LOAD_BALANCERS
+        spinnakerKind() >> SpinnakerKind.LOAD_BALANCERS
         kind() >> KubernetesKind.SERVICE
       }
     ])
 
     then:
-    kindMap.translateSpinnakerKind(KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES) == ImmutableSet.of(KubernetesKind.REPLICA_SET, KubernetesKind.DEPLOYMENT)
-    kindMap.translateSpinnakerKind(KubernetesSpinnakerKindMap.SpinnakerKind.LOAD_BALANCERS) == ImmutableSet.of(KubernetesKind.SERVICE)
-    kindMap.translateKubernetesKind(KubernetesKind.REPLICA_SET) == KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
-    kindMap.translateKubernetesKind(KubernetesKind.DEPLOYMENT) == KubernetesSpinnakerKindMap.SpinnakerKind.INSTANCES
-    kindMap.translateKubernetesKind(KubernetesKind.SERVICE) == KubernetesSpinnakerKindMap.SpinnakerKind.LOAD_BALANCERS
+    kindMap.translateSpinnakerKind(SpinnakerKind.INSTANCES) == ImmutableSet.of(KubernetesKind.REPLICA_SET, KubernetesKind.DEPLOYMENT)
+    kindMap.translateSpinnakerKind(SpinnakerKind.LOAD_BALANCERS) == ImmutableSet.of(KubernetesKind.SERVICE)
+    kindMap.translateKubernetesKind(KubernetesKind.REPLICA_SET) == SpinnakerKind.INSTANCES
+    kindMap.translateKubernetesKind(KubernetesKind.DEPLOYMENT) == SpinnakerKind.INSTANCES
+    kindMap.translateKubernetesKind(KubernetesKind.SERVICE) == SpinnakerKind.LOAD_BALANCERS
   }
 }
